[HLS-539] Fix compatibility between blocked GT transformer and WGR (#6)
* WIP


* existing tests pass


* rename file


* Add compat test


* scalafmt


* collect minimal columns


* address comments


* Test fixup


* Spark 3 needs more recent PyArrow, reduce mem consumption by removing unnecessary caching


* PyArrow 0.15.1 only with PySpark 3


* Don't use toPandas()


* Upgrade pyarrow


* Only register once


* Minimize memory usage


* Select before head


* set up/tear down


* Try limiting pyspark memory


* No teardown


* Extend timeout

+++ core/src/main/scala/io/projectglow/transformers/blockvariantsandsamples/VariantSampleBlockMaker.scala
@@ -38,7 +38,7 @@ private[projectglow] object VariantSampleBlockMaker extends GlowLogging {
     variantDf
       .withColumn(
         sortKeyField.name,
-        col(startField.name)
+        col(startField.name).cast(IntegerType)
       )
       .withColumn(
         headerField.name,
